bot/requirements.txt
python-telegram-bot[ext]>=21.0,<22.0
opensearch-py>=2.4,<3.0
httpx>=0.27,<1.0
